Fix bugs in pipeline plugin command
