+++ python/michelangelo/cli/mactl/plugins/trigger_run/main.py
@@ -3,8 +3,8 @@ from types import MethodType
 
 from grpc import Channel
 
-from mactl import CRD
-from plugins.trigger_run.kill import generate_kill
+from michelangelo.cli.mactl.crd import CRD
+from plugins.trigger_run.kill import generate_kill, add_function_signature
 
 
 _LOG = getLogger(__name__)
@@ -18,6 +18,7 @@ def apply_plugins(
     """
     _LOG.info("Applying plugins to crd: %r / %r", crd, target_command)
     if target_command == "kill":
+        add_function_signature(crd)
         crd.generate_kill = MethodType(
             lambda self, ch, parser: generate_kill(self, ch, parser), crd
         )
